Fixed a bug in GenReelsMainSymbolsDistance
